+++ src/ui/map_tests/map_feature_state.test.ts
@@ -1,4 +1,4 @@
-import {createMap, beforeMapTest, createStyleSource} from '../../../src/util/test/util';
+import {createMap, beforeMapTest, createStyleSource} from '../../util/test/util';
 
 beforeEach(() => {
     beforeMapTest();
