+++ src/ui/map_tests/map_is_moving.test.ts
@@ -1,8 +1,8 @@
-import {browser} from '../../../src/util/browser';
-import {Map} from '../../../src/ui/map';
-import {DOM} from '../../../src/util/dom';
-import simulate from '../../unit/lib/simulate_interaction';
-import {beforeMapTest} from '../../../src/util/test/util';
+import {browser} from '../../util/browser';
+import {Map} from '../map';
+import {DOM} from '../../util/dom';
+import simulate from '../../../test/unit/lib/simulate_interaction';
+import {beforeMapTest} from '../../util/test/util';
 
 let map;
