+++ src/ui/map_tests/map_style.test.ts
@@ -1,10 +1,10 @@
-import {Map, MapOptions} from '../../../src/ui/map';
-import {createMap, beforeMapTest, createStyle, createStyleSource} from '../../../src/util/test/util';
-import {Event as EventedEvent} from '../../../src/util/evented';
-import {fixedLngLat, fixedNum} from '../../unit/lib/fixed';
-import {extend} from '../../../src/util/util';
+import {Map, MapOptions} from '../map';
+import {createMap, beforeMapTest, createStyle, createStyleSource} from '../../util/test/util';
+import {Event as EventedEvent} from '../../util/evented';
+import {fixedLngLat, fixedNum} from '../../../test/unit/lib/fixed';
+import {extend} from '../../util/util';
 import {fakeServer, FakeServer} from 'nise';
-import {Style} from '../../../src/style/style';
+import {Style} from '../../style/style';
 import {GeoJSONSourceSpecification, LayerSpecification} from '@maplibre/maplibre-gl-style-spec';
 
 let server: FakeServer;
